extended class
